Added Lux Reading and Calibrated Utilizing Interpolation
