--- main.c
@@ -1,378 +1,0 @@
-#include "main.h"
-#include "ssd1306.h"
-#include "ssd1306_fonts.h"
-#include <string.h>
-#define UART_BUFFER_SIZE 128
-char uart_buffer[UART_BUFFER_SIZE];
-#include <stdlib.h>
-char light_value_str[32];
-char percentage_str[10];
-
-I2C_HandleTypeDef hi2c1;
-UART_HandleTypeDef huart1;
-ADC_HandleTypeDef hadc1;
-void SystemClock_Config(void);
-static void MX_GPIO_Init(void);
-static void MX_ADC1_Init(void);
-static void MX_I2C1_Init(void);
-static void MX_USART1_UART_Init(void);
-void Read_Light_Sensor(void);
-void FloatToString(float value, char *str);
-
-int main(void)
-{
-
-  HAL_Init();
-
-  SystemClock_Config();
-
-  MX_GPIO_Init();
-  MX_ADC1_Init();
-  MX_I2C1_Init();
-  MX_USART1_UART_Init();
-
-  ssd1306_Init();
-  		    ssd1306_Fill(Black);
-  		    ssd1306_UpdateScreen();
-  		    // Set up the OLED cursor
-  		    ssd1306_SetCursor(0, 0);
-  		    // Clear the UART buffer
-  		    memset(uart_buffer, 0, UART_BUFFER_SIZE);
-  		    // Display a startup message
-  		    ssd1306_WriteString("Waiting for data...", Font_6x8, White);
-  		    ssd1306_UpdateScreen();
-
-
-
-
-  while (1)
-  {
-	  if (HAL_UART_Receive(&huart1, (uint8_t *)uart_buffer, UART_BUFFER_SIZE - 1, HAL_MAX_DELAY) == HAL_OK) {
-	  		            // Ensure null termination
-	  		            uart_buffer[UART_BUFFER_SIZE - 1] = '\0';
-	  		          int uart_value = atoi(uart_buffer);  // Convert the string to an integer (if uart_buffer is a string)
-	  		          int percentage = (uart_value * 100) / 22600;
-	  		          itoa(percentage, percentage_str, 10);
-
-
-	  		            // Clear the display
-	  		            ssd1306_Fill(Black);
-	  		            ssd1306_SetCursor(0, 0);
-	  		            Read_Light_Sensor();
-	  		            // Display the received data
-	  		            ssd1306_WriteString("Plant Alarm v1.1", Font_6x8, White);
-	  		            ssd1306_SetCursor(0, 20); // Move cursor to the next line
-	  		            ssd1306_WriteString("Hydration(%):", Font_6x8, White);
-	  		            ssd1306_WriteString(percentage_str, Font_6x8, White);
-	  		            ssd1306_SetCursor(0, 40); // Move cursor to the next line
-	  		            ssd1306_WriteString("Lux:", Font_6x8, White);
-	  		            Read_Light_Sensor();
-	  		            ssd1306_WriteString(light_value_str, Font_6x8, White);
-
-	  		            // Update the screen
-	  		            ssd1306_UpdateScreen();
-	  		            HAL_Delay(1000);
-	  		            int moisture_value = atoi(uart_buffer);
-	  		                        // Control PA1 based on moisture value
-	  		                        if (moisture_value < 11000) {
-	  		                            HAL_GPIO_WritePin(GPIOA, GPIO_PIN_1, GPIO_PIN_SET); // Set PA1 high
-	  		                            ssd1306_Fill(Black);
-	  		                            ssd1306_SetCursor(0, 0);
-	  		                            ssd1306_WriteString("Please Water Plant! ", Font_6x8, White);
-	  		                            ssd1306_UpdateScreen();
-	  		                            HAL_Delay(500);
-	  		                          HAL_GPIO_WritePin(GPIOA, GPIO_PIN_1, GPIO_PIN_RESET);
-	  		                        } else {
-	  		                            HAL_GPIO_WritePin(GPIOA, GPIO_PIN_1, GPIO_PIN_RESET); // Set PA1 low
-	  		                        }
-	  		            // Clear the UART buffer for the next data
-	  		            memset(uart_buffer, 0, UART_BUFFER_SIZE);
-	  		        }
-  }
-
-}
-
-
-
-void FloatToString(float value, char *str) {
-    int int_part = (int)value;  // Get the integer part
-    int frac_part = (int)((value - int_part) * 100);  // Get the fractional part (2 decimal places)
-
-    // Convert integer part to string
-    int i = 0;
-    while (int_part > 0) {
-        str[i++] = '0' + (int_part % 10);
-        int_part /= 10;
-    }
-
-    // Reverse the integer part string
-    for (int j = 0; j < i / 2; j++) {
-        char temp = str[j];
-        str[j] = str[i - j - 1];
-        str[i - j - 1] = temp;
-    }
-
-    // Add the decimal point
-    str[i++] = '.';
-
-    // Convert fractional part to string (2 decimal places)
-    if (frac_part < 10) {
-        str[i++] = '0';  // Add leading zero for single digit fractional part
-    }
-    str[i++] = '0' + (frac_part / 10);  // Tens place
-    str[i++] = '0' + (frac_part % 10);  // Ones place
-
-    str[i] = '\0';  // Null-terminate the string
-}
-
-void Read_Light_Sensor(void) {
-    HAL_ADC_Start(&hadc1); // Start ADC conversion
-    if (HAL_ADC_PollForConversion(&hadc1, 100) == HAL_OK) {
-        uint32_t adc_value = HAL_ADC_GetValue(&hadc1); // Get the ADC value
-        float light_value = (adc_value * 3.3f) / 4095.0f; // Convert ADC value to voltage (assuming 12-bit ADC and 3.3V reference)
-
-        light_value= 10000.0f * exp(-3.684 * light_value);
-
-        // Convert the float light_value to string
-        FloatToString(light_value, light_value_str);
-
-        // Transmit the string via UART
-        HAL_UART_Transmit(&huart1, (uint8_t *)light_value_str, strlen(light_value_str), HAL_MAX_DELAY);
-    }
-    HAL_ADC_Stop(&hadc1); // Stop ADC conversion
-}
-
-
-void SystemClock_Config(void)
-{
-  RCC_OscInitTypeDef RCC_OscInitStruct = {0};
-  RCC_ClkInitTypeDef RCC_ClkInitStruct = {0};
-
-  /** Configure the main internal regulator output voltage
-  */
-  __HAL_RCC_PWR_CLK_ENABLE();
-  __HAL_PWR_VOLTAGESCALING_CONFIG(PWR_REGULATOR_VOLTAGE_SCALE2);
-
-  /** Initializes the RCC Oscillators according to the specified parameters
-  * in the RCC_OscInitTypeDef structure.
-  */
-  RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_HSI;
-  RCC_OscInitStruct.HSIState = RCC_HSI_ON;
-  RCC_OscInitStruct.HSICalibrationValue = RCC_HSICALIBRATION_DEFAULT;
-  RCC_OscInitStruct.PLL.PLLState = RCC_PLL_ON;
-  RCC_OscInitStruct.PLL.PLLSource = RCC_PLLSOURCE_HSI;
-  RCC_OscInitStruct.PLL.PLLM = 16;
-  RCC_OscInitStruct.PLL.PLLN = 336;
-  RCC_OscInitStruct.PLL.PLLP = RCC_PLLP_DIV4;
-  RCC_OscInitStruct.PLL.PLLQ = 7;
-  if (HAL_RCC_OscConfig(&RCC_OscInitStruct) != HAL_OK)
-  {
-    Error_Handler();
-  }
-
-  /** Initializes the CPU, AHB and APB buses clocks
-  */
-  RCC_ClkInitStruct.ClockType = RCC_CLOCKTYPE_HCLK|RCC_CLOCKTYPE_SYSCLK
-                              |RCC_CLOCKTYPE_PCLK1|RCC_CLOCKTYPE_PCLK2;
-  RCC_ClkInitStruct.SYSCLKSource = RCC_SYSCLKSOURCE_PLLCLK;
-  RCC_ClkInitStruct.AHBCLKDivider = RCC_SYSCLK_DIV1;
-  RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV2;
-  RCC_ClkInitStruct.APB2CLKDivider = RCC_HCLK_DIV1;
-
-  if (HAL_RCC_ClockConfig(&RCC_ClkInitStruct, FLASH_LATENCY_2) != HAL_OK)
-  {
-    Error_Handler();
-  }
-}
-
-/**
-  * @brief ADC1 Initialization Function
-  * @param None
-  * @retval None
-  */
-static void MX_ADC1_Init(void)
-{
-
-  /* USER CODE BEGIN ADC1_Init 0 */
-
-  /* USER CODE END ADC1_Init 0 */
-
-  ADC_ChannelConfTypeDef sConfig = {0};
-
-  /* USER CODE BEGIN ADC1_Init 1 */
-
-  /* USER CODE END ADC1_Init 1 */
-
-  /** Configure the global features of the ADC (Clock, Resolution, Data Alignment and number of conversion)
-  */
-  hadc1.Instance = ADC1;
-  hadc1.Init.ClockPrescaler = ADC_CLOCK_SYNC_PCLK_DIV4;
-  hadc1.Init.Resolution = ADC_RESOLUTION_12B;
-  hadc1.Init.ScanConvMode = DISABLE;
-  hadc1.Init.ContinuousConvMode = DISABLE;
-  hadc1.Init.DiscontinuousConvMode = DISABLE;
-  hadc1.Init.ExternalTrigConvEdge = ADC_EXTERNALTRIGCONVEDGE_NONE;
-  hadc1.Init.ExternalTrigConv = ADC_SOFTWARE_START;
-  hadc1.Init.DataAlign = ADC_DATAALIGN_RIGHT;
-  hadc1.Init.NbrOfConversion = 1;
-  hadc1.Init.DMAContinuousRequests = DISABLE;
-  hadc1.Init.EOCSelection = ADC_EOC_SINGLE_CONV;
-  if (HAL_ADC_Init(&hadc1) != HAL_OK)
-  {
-    Error_Handler();
-  }
-
-  /** Configure for the selected ADC regular channel its corresponding rank in the sequencer and its sample time.
-  */
-  sConfig.Channel = ADC_CHANNEL_4;
-  sConfig.Rank = 1;
-  sConfig.SamplingTime = ADC_SAMPLETIME_3CYCLES;
-  if (HAL_ADC_ConfigChannel(&hadc1, &sConfig) != HAL_OK)
-  {
-    Error_Handler();
-  }
-  /* USER CODE BEGIN ADC1_Init 2 */
-
-  /* USER CODE END ADC1_Init 2 */
-
-}
-
-/**
-  * @brief I2C1 Initialization Function
-  * @param None
-  * @retval None
-  */
-static void MX_I2C1_Init(void)
-{
-
-  /* USER CODE BEGIN I2C1_Init 0 */
-
-  /* USER CODE END I2C1_Init 0 */
-
-  /* USER CODE BEGIN I2C1_Init 1 */
-
-  /* USER CODE END I2C1_Init 1 */
-  hi2c1.Instance = I2C1;
-  hi2c1.Init.ClockSpeed = 100000;
-  hi2c1.Init.DutyCycle = I2C_DUTYCYCLE_2;
-  hi2c1.Init.OwnAddress1 = 0;
-  hi2c1.Init.AddressingMode = I2C_ADDRESSINGMODE_7BIT;
-  hi2c1.Init.DualAddressMode = I2C_DUALADDRESS_DISABLE;
-  hi2c1.Init.OwnAddress2 = 0;
-  hi2c1.Init.GeneralCallMode = I2C_GENERALCALL_DISABLE;
-  hi2c1.Init.NoStretchMode = I2C_NOSTRETCH_DISABLE;
-  if (HAL_I2C_Init(&hi2c1) != HAL_OK)
-  {
-    Error_Handler();
-  }
-  /* USER CODE BEGIN I2C1_Init 2 */
-
-  /* USER CODE END I2C1_Init 2 */
-
-}
-
-/**
-  * @brief USART1 Initialization Function
-  * @param None
-  * @retval None
-  */
-static void MX_USART1_UART_Init(void)
-{
-
-  /* USER CODE BEGIN USART1_Init 0 */
-
-  /* USER CODE END USART1_Init 0 */
-
-  /* USER CODE BEGIN USART1_Init 1 */
-
-  /* USER CODE END USART1_Init 1 */
-  huart1.Instance = USART1;
-  huart1.Init.BaudRate = 115200;
-  huart1.Init.WordLength = UART_WORDLENGTH_8B;
-  huart1.Init.StopBits = UART_STOPBITS_1;
-  huart1.Init.Parity = UART_PARITY_NONE;
-  huart1.Init.Mode = UART_MODE_TX_RX;
-  huart1.Init.HwFlowCtl = UART_HWCONTROL_NONE;
-  huart1.Init.OverSampling = UART_OVERSAMPLING_16;
-  if (HAL_UART_Init(&huart1) != HAL_OK)
-  {
-    Error_Handler();
-  }
-  /* USER CODE BEGIN USART1_Init 2 */
-
-  /* USER CODE END USART1_Init 2 */
-
-}
-
-/**
-  * @brief GPIO Initialization Function
-  * @param None
-  * @retval None
-  */
-static void MX_GPIO_Init(void)
-{
-  GPIO_InitTypeDef GPIO_InitStruct = {0};
-/* USER CODE BEGIN MX_GPIO_Init_1 */
-/* USER CODE END MX_GPIO_Init_1 */
-
-  /* GPIO Ports Clock Enable */
-  __HAL_RCC_GPIOC_CLK_ENABLE();
-  __HAL_RCC_GPIOH_CLK_ENABLE();
-  __HAL_RCC_GPIOA_CLK_ENABLE();
-  __HAL_RCC_GPIOB_CLK_ENABLE();
-
-  /*Configure GPIO pin Output Level */
-  HAL_GPIO_WritePin(GPIOA, GPIO_PIN_1|LD2_Pin, GPIO_PIN_RESET);
-
-  /*Configure GPIO pin : B1_Pin */
-  GPIO_InitStruct.Pin = B1_Pin;
-  GPIO_InitStruct.Mode = GPIO_MODE_IT_FALLING;
-  GPIO_InitStruct.Pull = GPIO_NOPULL;
-  HAL_GPIO_Init(B1_GPIO_Port, &GPIO_InitStruct);
-
-  /*Configure GPIO pins : PA1 LD2_Pin */
-  GPIO_InitStruct.Pin = GPIO_PIN_1|LD2_Pin;
-  GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-  GPIO_InitStruct.Pull = GPIO_NOPULL;
-  GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-  HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
-
-/* USER CODE BEGIN MX_GPIO_Init_2 */
-/* USER CODE END MX_GPIO_Init_2 */
-}
-
-/* USER CODE BEGIN 4 */
-
-/* USER CODE END 4 */
-
-/**
-  * @brief  This function is executed in case of error occurrence.
-  * @retval None
-  */
-void Error_Handler(void)
-{
-  /* USER CODE BEGIN Error_Handler_Debug */
-  /* User can add his own implementation to report the HAL error return state */
-  __disable_irq();
-  while (1)
-  {
-  }
-  /* USER CODE END Error_Handler_Debug */
-}
-
-#ifdef  USE_FULL_ASSERT
-/**
-  * @brief  Reports the name of the source file and the source line number
-  *         where the assert_param error has occurred.
-  * @param  file: pointer to the source file name
-  * @param  line: assert_param error line source number
-  * @retval None
-  */
-void assert_failed(uint8_t *file, uint32_t line)
-{
-  /* USER CODE BEGIN 6 */
-  /* User can add his own implementation to report the file name and line number,
-     ex: printf("Wrong parameters value: file %s on line %d\r\n", file, line) */
-  /* USER CODE END 6 */
-}
-#endif /* USE_FULL_ASSERT */
